(WIP) fix KotlinGradleSensor
