+++ sonar-kotlin-api/src/test/java/org/sonarsource/kotlin/api/sensors/AbstractKotlinSensorTest.kt
@@ -262,6 +262,7 @@ class DummyKotlinSensor(checkFactory: CheckFactory, language: KotlinLanguage, ch
         sensorContext, filesToAnalyze, progressReport, listOf(KtChecksVisitor(checks)), filenames, LOG
     ) {
         override val bindingContext: BindingContext = BindingContext.EMPTY
+        override val doResolve: Boolean = false
     }
 
     override fun getFilesToAnalyse(sensorContext: SensorContext): Iterable<InputFile> =
