+++ sonar-kotlin-checks/src/main/java/org/sonarsource/kotlin/checks/IgnoredOperationStatusCheck.kt
@@ -32,7 +32,7 @@ import org.sonarsource.kotlin.api.checks.CallAbstractCheck
 import org.sonarsource.kotlin.api.checks.FunMatcher
 import org.sonarsource.kotlin.api.checks.simpleName
 import org.sonarsource.kotlin.api.frontend.KotlinFileContext
-import org.sonarsource.kotlin.visiting.analyze
+import org.sonarsource.kotlin.api.visiting.analyze
 
 @org.sonarsource.kotlin.api.frontend.K1only("easy?")
 @Rule(key = "S899")
