+++ sonar-kotlin-plugin/src/main/java/org/sonarsource/kotlin/plugin/KotlinSensor.kt
@@ -89,6 +89,8 @@ class KotlinSensor(
                 BindingContext.EMPTY
             }
         }
+
+        override val doResolve: Boolean = true
     }
 
     private fun visitors(sensorContext: SensorContext): List<KotlinFileVisitor> =
