+++ sonar-kotlin-test-api/src/main/java/org/sonarsource/kotlin/testapi/KtTestChecksVisitor.kt
@@ -29,7 +29,6 @@ import org.sonar.api.rule.RuleKey
 import org.sonarsource.kotlin.api.checks.AbstractCheck
 import org.sonarsource.kotlin.api.frontend.KotlinFileContext
 import org.sonarsource.kotlin.api.visiting.KotlinFileVisitor
-import org.sonarsource.kotlin.visiting.kaSession
 
 class KtTestChecksVisitor(private val checks: List<AbstractCheck>) : KotlinFileVisitor() {
     init {
@@ -45,7 +44,7 @@ class KtTestChecksVisitor(private val checks: List<AbstractCheck>) : KotlinFileV
         }
     }
 
-    override fun visit(kotlinFileContext: KotlinFileContext) = kaSession(kotlinFileContext.ktFile) {
+    override fun visit(kotlinFileContext: KotlinFileContext) {
         flattenNodes(listOf(kotlinFileContext.ktFile)).forEach { psiElement ->
             // Note: we only visit KtElements. If we need to visit PsiElement, add a
             // visitPsiElement function in KotlinCheck and call it here in the else branch.
